Added registration

+++ registration.py
@@ -189,7 +189,7 @@ def resample(
         align_corners=True
     )
     moved = func.grid_sample(
-        image_tensor, grid,
+        image_tensor.to(torch.float32), grid,
         align_corners=True
     ).view(output_dims.shape)
